amibec_src/requirements.txt
asgiref==3.4.1
Babel==2.9.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
defusedxml==0.7.1
Django==2.2.24
django-allauth==0.46.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-extra-views==0.13.0
django-haystack==3.1.1
django-localflavor==3.1
django-oscar==3.1
django-oscar-paypal==2.0.0
django-phonenumber-field==5.2.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.9
factory-boy==3.2.1
Faker==9.8.2
idna==3.3
oauthlib==3.1.1
phonenumbers==8.12.37
Pillow==8.4.0
purl==1.6
pycountry==20.7.3
pycparser==2.21
PyJWT==2.3.0
pysolr==3.9.0
python-dateutil==2.8.2
python-stdnum==1.17
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
